Moving Settings to Seperate Project
Edited SimpleInjector packages and references to IDukeboxSettings, config is now in seperate XML config file 'Dukebox.config'
