+++ Dukebox.Audio/Dukebox.Audio.csproj
@@ -130,7 +130,12 @@
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
     </None>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\Dukebox.Configuration\Dukebox.Configuration.csproj">
+      <Project>{b932e6d9-1e93-47c4-a84e-30e0e1d4e4d7}</Project>
+      <Name>Dukebox.Configuration</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
