+++ Dukebox.Configuration/Dukebox.Config
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <appSettings>
+    <add key="albumArtCachePath" value="%APPDATA%/Dukebox/albumArtCache"/>
+    <add key="addDirectoryConcurrencyLimit" value="10"/>
+    <add key="bassPluginCount" value="14"/>
+    <add key="bassLicenseEmail" value="WkdwbVpIbDFjblZwY25sQWFHOTBiV0ZwYkM1amIyMD0="/>
+    <add key="bassLicenseKey" value="TWxnNU1Ua3pOREl6TXpFeU5ESXk="/>
+    <add key="bassAddOnsPath" value=".\BASS\"/>
+    <add key="trackDisplayFormat" value="{artist} - {title}"/>
+  </appSettings>
+</configuration>
