+++ Dukebox.Configuration/Interfaces/IDukeboxSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace Dukebox.Configuration.Interfaces
+{
+    public interface IDukeboxSettings
+    {
+        int AddDirectoryConcurrencyLimit { get; }
+        string AlbumArtCachePath { get; }
+        int BassPluginCount { get; }
+        string BassLicenseEmail { get; }
+        string BassLicenseKey { get; }
+        string BassAddOnsPath { get; }
+        string TrackDisplayFormat { get; }
+    }
+}
