+++ Dukebox.Configuration/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="SimpleInjector" version="3.1.2" targetFramework="net45" />
+  <package id="SimpleInjector.Packaging" version="3.1.2" targetFramework="net45" />
+</packages>
