+++ Dukebox.Desktop/app.config
@@ -14,15 +14,6 @@
   <connectionStrings>
     <add name="Library" connectionString="data source=&quot;|DataDirectory|\library.s3db&quot;" providerName="System.Data.SQLite"/>
   </connectionStrings>
-  <appSettings>
-    <add key="albumArtCachePath" value="%APPDATA%/Dukebox/albumArtCache"/>
-    <add key="addDirectoryConcurrencyLimit" value="10"/>
-    <add key="bassPluginCount" value="14"/>
-    <add key="bassLicenseEmail" value="WkdwbVpIbDFjblZwY25sQWFHOTBiV0ZwYkM1amIyMD0="/>
-    <add key="bassLicenseKey" value="TWxnNU1Ua3pOREl6TXpFeU5ESXk="/>
-    <add key="bassAddOnsPath" value=".\BASS\"/>
-    <add key="trackDisplayFormat" value="{artist} - {title}"/>
-  </appSettings>
   <userSettings>
     <Dukebox.Desktop.Properties.Settings>
       <setting name="shuffle" serializeAs="String">
