+++ Dukebox.Library/Dukebox.Library.csproj
@@ -137,14 +137,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Factories\TrackFactory.cs" />
-    <Compile Include="Helper\Base64Decoder.cs" />
     <Compile Include="Model\Album.cs" />
     <Compile Include="Model\Artist.cs" />
     <Compile Include="Helper\DukeboxInitialisationHelper.cs" />
     <Compile Include="Interfaces\IAudioCdDriveMonitoringService.cs" />
     <Compile Include="Interfaces\ICdRipViewUpdater.cs" />
-    <Compile Include="Config\DukeboxSettings.cs" />
-    <Compile Include="Interfaces\IDukeboxSettings.cs" />
     <Compile Include="Interfaces\ITrack.cs" />
     <Compile Include="Interfaces\IMusicLibraryDbContext.cs" />
     <Compile Include="Repositories\MusicLibraryDbContext.cs" />
@@ -188,6 +185,10 @@
     <ProjectReference Include="..\Dukebox.Audio\Dukebox.Audio.csproj">
       <Project>{8ea07179-2c2a-42a3-90c4-c700c4681181}</Project>
       <Name>Dukebox.Audio</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Dukebox.Configuration\Dukebox.Configuration.csproj">
+      <Project>{b932e6d9-1e93-47c4-a84e-30e0e1d4e4d7}</Project>
+      <Name>Dukebox.Configuration</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
