+++ Dukebox.Tests/Dukebox.Tests.csproj
@@ -148,9 +148,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
-    <None Include="Dukebox.exe.Config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
@@ -161,6 +158,10 @@
     <ProjectReference Include="..\Dukebox.Audio\Dukebox.Audio.csproj">
       <Project>{8ea07179-2c2a-42a3-90c4-c700c4681181}</Project>
       <Name>Dukebox.Audio</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Dukebox.Configuration\Dukebox.Configuration.csproj">
+      <Project>{b932e6d9-1e93-47c4-a84e-30e0e1d4e4d7}</Project>
+      <Name>Dukebox.Configuration</Name>
     </ProjectReference>
     <ProjectReference Include="..\Dukebox.Library\Dukebox.Library.csproj">
       <Project>{6c5b8890-79a4-43ee-9734-b5db714fcf16}</Project>
