--- Dukebox.Tests/Dukebox.exe.Config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0"?>
-<configuration>
-  <appSettings>
-    <add key="albumArtCachePath" value="./albumArtCache"/>
-    <add key="addDirectoryConcurrencyLimit" value="10"/>
-    <add key="bassPluginCount" value="13"/>
-    <add key="bassLicenseEmail" value="WkdwbVpIbDFjblZwY25sQWFHOTBiV0ZwYkM1amIyMD0="/>
-    <add key="bassLicenseKey" value="TWxnNU1Ua3pOREl6TXpFeU5ESXk="/>
-    <add key="bassAddOnsPath" value="M:\Source Code\C#\Dukebox\code\Dukebox.Tests\bin\Debug\BASS\"/>
-    <add key="trackDisplayFormat" value="{artist} - {title}"/>
-  </appSettings>
-</configuration>
