+++ Dukebox.sln
@@ -28,62 +28,102 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{DF959C36-6
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Libraries", "Libraries", "{430343AB-8209-4B92-B9BC-53F8632657A9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dukebox.Configuration", "Dukebox.Configuration\Dukebox.Configuration.csproj", "{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|x64.ActiveCfg = Debug|x64
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|x64.Build.0 = Debug|x64
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|x86.ActiveCfg = Debug|x86
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Debug|x86.Build.0 = Debug|x86
+		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|x64.ActiveCfg = Release|x64
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|x64.Build.0 = Release|x64
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|x86.ActiveCfg = Release|x86
 		{8EA07179-2C2A-42A3-90C4-C700C4681181}.Release|x86.Build.0 = Release|x86
+		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|x64.ActiveCfg = Debug|x64
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|x64.Build.0 = Debug|x64
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|x86.ActiveCfg = Debug|x86
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Debug|x86.Build.0 = Debug|x86
+		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|x64.ActiveCfg = Release|x64
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|x64.Build.0 = Release|x64
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|x86.ActiveCfg = Release|x86
 		{6C5B8890-79A4-43EE-9734-B5DB714FCF16}.Release|x86.Build.0 = Release|x86
+		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|x64.ActiveCfg = Debug|x64
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|x64.Build.0 = Debug|x64
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|x86.ActiveCfg = Debug|x86
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Debug|x86.Build.0 = Debug|x86
+		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|x64.ActiveCfg = Release|x64
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|x64.Build.0 = Release|x64
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|x86.ActiveCfg = Release|x86
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99}.Release|x86.Build.0 = Release|x86
+		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|x64.ActiveCfg = Debug|x64
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|x64.Build.0 = Debug|x64
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|x86.ActiveCfg = Debug|x86
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Debug|x86.Build.0 = Debug|x86
+		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|x64.ActiveCfg = Release|x64
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|x64.Build.0 = Release|x64
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|x86.ActiveCfg = Release|x86
 		{2CE69F0E-15ED-47AE-A285-4197D458443F}.Release|x86.Build.0 = Release|x86
+		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|x64.ActiveCfg = Debug|x64
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|x64.Build.0 = Debug|x64
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|x86.ActiveCfg = Debug|x86
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Debug|x86.Build.0 = Debug|x86
+		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|x64.ActiveCfg = Release|x64
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|x64.Build.0 = Release|x64
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|x86.ActiveCfg = Release|x86
 		{5BE6D473-B49E-451C-94F9-19CA55997030}.Release|x86.Build.0 = Release|x86
+		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|x64.ActiveCfg = Debug|x64
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|x64.Build.0 = Debug|x64
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|x86.ActiveCfg = Debug|x86
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Debug|x86.Build.0 = Debug|x86
+		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|Any CPU.Build.0 = Release|Any CPU
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|x64.ActiveCfg = Release|x64
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|x64.Build.0 = Release|x64
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|x86.ActiveCfg = Release|x86
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351}.Release|x86.Build.0 = Release|x86
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|x64.ActiveCfg = Debug|x64
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|x64.Build.0 = Debug|x64
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|x86.ActiveCfg = Debug|x86
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Debug|x86.Build.0 = Debug|x86
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|x64.ActiveCfg = Release|x64
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|x64.Build.0 = Release|x64
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|x86.ActiveCfg = Release|x86
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -94,5 +134,6 @@ Global
 		{E69D9EB5-0707-41EE-8537-3FFEA9A8BB99} = {430343AB-8209-4B92-B9BC-53F8632657A9}
 		{5BE6D473-B49E-451C-94F9-19CA55997030} = {DF959C36-67C4-4013-A831-1B205D6C8B7C}
 		{07AD0C1B-27B5-478B-B87F-610F3CBEC351} = {8982A846-7F4B-499D-8CE7-5C5A2C38F7ED}
+		{B932E6D9-1E93-47C4-A84E-30E0E1D4E4D7} = {430343AB-8209-4B92-B9BC-53F8632657A9}
 	EndGlobalSection
 EndGlobal
